requirements.txt
ffmpeg-python
scenedetect
opencv-python
imageio
imageio-ffmpeg